refactor(built-in-ai): drop the stale error demand, file the store as a store
EngineSnapshot kept requiring "error" after c9cbab4 deleted its only
reader (the MissingUserActivationError check the 0.8.0 downloadable
status made obsolete). The Pick now states what deriveEngineView reads.

use-ai-shared-context.ts was a persisted store named like a hook;
renamed to shared-context-store.ts to match language/playback/speech
store siblings. Exports unchanged.
